requirements.txt
pandas==2.2.1
boto3==1.34.64
jpholiday==0.1.8
workdays==1.4
slack-bolt==1.18.1
requests==2.31.0
python-dotenv==1.0.1